Revert "2nd Comming in the rebuild"
This reverts commit c1e559969a482ba3c3c4b56c1f84959882ae0243.

+++ W-SmartShopSelution/WPF GUI/Orders/In/IncomeOrder/IncomeOrderUC.xaml.cs
@@ -954,7 +954,7 @@ namespace WPF_GUI
             }*/
             foreach(IncomeOrderProductModel incomeOrderProduct in IncomeOrderProducts)
             {
-                totalPrice += incomeOrderProduct.Product.IncomePrice * incomeOrderProduct.Quantity;
+                totalPrice += incomeOrderProduct.Product.IncomePrice * (decimal)incomeOrderProduct.Quantity;
             }
             TotalPriceValue_IncomeOrderUC.Text = totalPrice.ToString();
